feat: support nested heroines if they are not part of the same transition
